+++ Assets/SteamVR_Input/SteamVR_Input_Actions.cs
@@ -113,6 +113,14 @@ namespace Valve.VR
         
         private static SteamVR_Action_Vector2 p_cockpitControls_POV2JoystickPosition;
         
+        private static SteamVR_Action_Boolean p_fSSControls_ExitFSS;
+        
+        private static SteamVR_Action_Vector2 p_fSSControls_CameraControl;
+        
+        private static SteamVR_Action_Boolean p_fSSControls_CameraControlActivate;
+        
+        private static SteamVR_Action_Boolean p_fSSControls_TargetCurrentSignal;
+        
         public static SteamVR_Action_Boolean default_InteractUI
         {
             get
@@ -497,6 +505,38 @@ namespace Valve.VR
             }
         }
         
+        public static SteamVR_Action_Boolean fSSControls_ExitFSS
+        {
+            get
+            {
+                return SteamVR_Actions.p_fSSControls_ExitFSS.GetCopy<SteamVR_Action_Boolean>();
+            }
+        }
+        
+        public static SteamVR_Action_Vector2 fSSControls_CameraControl
+        {
+            get
+            {
+                return SteamVR_Actions.p_fSSControls_CameraControl.GetCopy<SteamVR_Action_Vector2>();
+            }
+        }
+        
+        public static SteamVR_Action_Boolean fSSControls_CameraControlActivate
+        {
+            get
+            {
+                return SteamVR_Actions.p_fSSControls_CameraControlActivate.GetCopy<SteamVR_Action_Boolean>();
+            }
+        }
+        
+        public static SteamVR_Action_Boolean fSSControls_TargetCurrentSignal
+        {
+            get
+            {
+                return SteamVR_Actions.p_fSSControls_TargetCurrentSignal.GetCopy<SteamVR_Action_Boolean>();
+            }
+        }
+        
         private static void InitializeActionArrays()
         {
             Valve.VR.SteamVR_Input.actions = new Valve.VR.SteamVR_Action[] {
@@ -547,7 +587,11 @@ namespace Valve.VR
                     SteamVR_Actions.cockpitControls_POV2TrackpadPress,
                     SteamVR_Actions.cockpitControls_POV2TrackpadTouch,
                     SteamVR_Actions.cockpitControls_POV2TrackpadPosition,
-                    SteamVR_Actions.cockpitControls_POV2JoystickPosition};
+                    SteamVR_Actions.cockpitControls_POV2JoystickPosition,
+                    SteamVR_Actions.fSSControls_ExitFSS,
+                    SteamVR_Actions.fSSControls_CameraControl,
+                    SteamVR_Actions.fSSControls_CameraControlActivate,
+                    SteamVR_Actions.fSSControls_TargetCurrentSignal};
             Valve.VR.SteamVR_Input.actionsIn = new Valve.VR.ISteamVR_Action_In[] {
                     SteamVR_Actions.default_InteractUI,
                     SteamVR_Actions.default_GrabToggle,
@@ -595,7 +639,11 @@ namespace Valve.VR
                     SteamVR_Actions.cockpitControls_POV2TrackpadPress,
                     SteamVR_Actions.cockpitControls_POV2TrackpadTouch,
                     SteamVR_Actions.cockpitControls_POV2TrackpadPosition,
-                    SteamVR_Actions.cockpitControls_POV2JoystickPosition};
+                    SteamVR_Actions.cockpitControls_POV2JoystickPosition,
+                    SteamVR_Actions.fSSControls_ExitFSS,
+                    SteamVR_Actions.fSSControls_CameraControl,
+                    SteamVR_Actions.fSSControls_CameraControlActivate,
+                    SteamVR_Actions.fSSControls_TargetCurrentSignal};
             Valve.VR.SteamVR_Input.actionsOut = new Valve.VR.ISteamVR_Action_Out[] {
                     SteamVR_Actions.default_Haptic};
             Valve.VR.SteamVR_Input.actionsVibration = new Valve.VR.SteamVR_Action_Vibration[] {
@@ -638,7 +686,10 @@ namespace Valve.VR
                     SteamVR_Actions.cockpitControls_POV1TrackpadTouch,
                     SteamVR_Actions.cockpitControls_ButtonPOV2,
                     SteamVR_Actions.cockpitControls_POV2TrackpadPress,
-                    SteamVR_Actions.cockpitControls_POV2TrackpadTouch};
+                    SteamVR_Actions.cockpitControls_POV2TrackpadTouch,
+                    SteamVR_Actions.fSSControls_ExitFSS,
+                    SteamVR_Actions.fSSControls_CameraControlActivate,
+                    SteamVR_Actions.fSSControls_TargetCurrentSignal};
             Valve.VR.SteamVR_Input.actionsSingle = new Valve.VR.SteamVR_Action_Single[0];
             Valve.VR.SteamVR_Input.actionsVector2 = new Valve.VR.SteamVR_Action_Vector2[] {
                     SteamVR_Actions.menu_MenuNavigateTrackpadPosition,
@@ -650,7 +701,8 @@ namespace Valve.VR
                     SteamVR_Actions.cockpitControls_POV1TrackpadPosition,
                     SteamVR_Actions.cockpitControls_POV1JoystickPosition,
                     SteamVR_Actions.cockpitControls_POV2TrackpadPosition,
-                    SteamVR_Actions.cockpitControls_POV2JoystickPosition};
+                    SteamVR_Actions.cockpitControls_POV2JoystickPosition,
+                    SteamVR_Actions.fSSControls_CameraControl};
             Valve.VR.SteamVR_Input.actionsVector3 = new Valve.VR.SteamVR_Action_Vector3[0];
             Valve.VR.SteamVR_Input.actionsSkeleton = new Valve.VR.SteamVR_Action_Skeleton[0];
             Valve.VR.SteamVR_Input.actionsNonPoseNonSkeletonIn = new Valve.VR.ISteamVR_Action_In[] {
@@ -699,7 +751,11 @@ namespace Valve.VR
                     SteamVR_Actions.cockpitControls_POV2TrackpadPress,
                     SteamVR_Actions.cockpitControls_POV2TrackpadTouch,
                     SteamVR_Actions.cockpitControls_POV2TrackpadPosition,
-                    SteamVR_Actions.cockpitControls_POV2JoystickPosition};
+                    SteamVR_Actions.cockpitControls_POV2JoystickPosition,
+                    SteamVR_Actions.fSSControls_ExitFSS,
+                    SteamVR_Actions.fSSControls_CameraControl,
+                    SteamVR_Actions.fSSControls_CameraControlActivate,
+                    SteamVR_Actions.fSSControls_TargetCurrentSignal};
         }
         
         private static void PreInitActions()
@@ -752,6 +808,10 @@ namespace Valve.VR
             SteamVR_Actions.p_cockpitControls_POV2TrackpadTouch = ((SteamVR_Action_Boolean)(SteamVR_Action.Create<SteamVR_Action_Boolean>("/actions/CockpitControls/in/POV2TrackpadTouch")));
             SteamVR_Actions.p_cockpitControls_POV2TrackpadPosition = ((SteamVR_Action_Vector2)(SteamVR_Action.Create<SteamVR_Action_Vector2>("/actions/CockpitControls/in/POV2TrackpadPosition")));
             SteamVR_Actions.p_cockpitControls_POV2JoystickPosition = ((SteamVR_Action_Vector2)(SteamVR_Action.Create<SteamVR_Action_Vector2>("/actions/CockpitControls/in/POV2JoystickPosition")));
+            SteamVR_Actions.p_fSSControls_ExitFSS = ((SteamVR_Action_Boolean)(SteamVR_Action.Create<SteamVR_Action_Boolean>("/actions/FSSControls/in/ExitFSS")));
+            SteamVR_Actions.p_fSSControls_CameraControl = ((SteamVR_Action_Vector2)(SteamVR_Action.Create<SteamVR_Action_Vector2>("/actions/FSSControls/in/CameraControl")));
+            SteamVR_Actions.p_fSSControls_CameraControlActivate = ((SteamVR_Action_Boolean)(SteamVR_Action.Create<SteamVR_Action_Boolean>("/actions/FSSControls/in/CameraControlActivate")));
+            SteamVR_Actions.p_fSSControls_TargetCurrentSignal = ((SteamVR_Action_Boolean)(SteamVR_Action.Create<SteamVR_Action_Boolean>("/actions/FSSControls/in/TargetCurrentSignal")));
         }
     }
 }